ADM-873:[frontend]refactor: refactor code
