+++ frontend/src/containers/ReportButtonGroup/index.tsx
@@ -6,9 +6,11 @@ import { COMMON_BUTTONS, REPORT_TYPES } from '@src/constants/commons';
 import { ReportResponseDTO } from '@src/clients/report/dto/response';
 import { useExportCsvEffect } from '@src/hooks/useExportCsvEffect';
 import SaveAltIcon from '@mui/icons-material/SaveAlt';
-import { TIPS } from '@src/constants/resources';
+import {MESSAGE, TIPS} from '@src/constants/resources';
 import { Tooltip } from '@mui/material';
-import React from 'react';
+import React, {useEffect, useLayoutEffect, useState} from 'react';
+import {useAppDispatch} from "@src/hooks/useAppDispatch";
+import {addNotification} from "@src/context/notification/NotificationSlice";
 
 interface ReportButtonGroupProps {
   handleSave?: () => void;
@@ -35,7 +37,10 @@ export const ReportButtonGroup = ({
   isShowExportBoardButton,
   isShowExportPipelineButton,
 }: ReportButtonGroupProps) => {
+  const dispatch = useAppDispatch();
   const { fetchExportData, isExpired } = useExportCsvEffect();
+  const [exportValidityTimeMin, setExportValidityTimeMin] = useState<number | undefined | null>(undefined);
+  const [allMetricsCompleted, setAllMetricsCompleted] = useState<boolean>(false);
 
   const exportCSV = (dataType: REPORT_TYPES, startDate: string, endDate: string): CSVReportRequestDTO => ({
     dataType: dataType,
@@ -59,6 +64,54 @@ export const ReportButtonGroup = ({
     !!reportData?.reportMetricsError.pipelineMetricsError ||
     !!reportData?.reportMetricsError.sourceControlMetricsError;
 
+  const enableExportMetricData = !!(reportData?.overallMetricsCompleted && !isReportHasError);
+  const enableExportBoardData = !!(
+    reportData?.boardMetricsCompleted && !reportData?.reportMetricsError?.boardMetricsError
+  );
+  const enableExportPipelineData = !pipelineButtonDisabled;
+
+  useEffect(() => {
+    setExportValidityTimeMin(reportData?.exportValidityTime);
+    reportData?.allMetricsCompleted &&
+    setAllMetricsCompleted(enableExportMetricData || enableExportBoardData || enableExportPipelineData);
+  }, [dispatch, reportData]);
+
+  useLayoutEffect(() => {
+    if (exportValidityTimeMin && allMetricsCompleted) {
+      const startTime = Date.now();
+      const timer = setInterval(() => {
+        const currentTime = Date.now();
+        const elapsedTime = currentTime - startTime;
+
+        const remainingExpireTime = 5 * 60 * 1000;
+        const remainingTime = exportValidityTimeMin * 60 * 1000 - elapsedTime;
+        if (remainingTime <= remainingExpireTime) {
+          dispatch(
+            addNotification({
+              message: MESSAGE.EXPIRE_INFORMATION(5),
+            }),
+          );
+          clearInterval(timer);
+        }
+      }, 1000);
+
+      return () => {
+        clearInterval(timer);
+      };
+    }
+  }, [dispatch, exportValidityTimeMin, allMetricsCompleted]);
+
+
+  useLayoutEffect(() => {
+    exportValidityTimeMin &&
+    allMetricsCompleted &&
+    dispatch(
+      addNotification({
+        message: MESSAGE.EXPIRE_INFORMATION(exportValidityTimeMin),
+      }),
+    );
+  }, [dispatch, exportValidityTimeMin, allMetricsCompleted]);
+
   return (
     <>
       <StyledButtonGroup isShowSave={isShowSave}>
@@ -75,7 +128,7 @@ export const ReportButtonGroup = ({
           </BackButton>
           {isShowExportMetrics && (
             <StyledExportButton
-              disabled={!(reportData?.overallMetricsCompleted && !isReportHasError)}
+              disabled={!enableExportMetricData}
               onClick={() => handleDownload(REPORT_TYPES.METRICS, startDate, endDate)}
             >
               {COMMON_BUTTONS.EXPORT_METRIC_DATA}
@@ -83,7 +136,7 @@ export const ReportButtonGroup = ({
           )}
           {isShowExportBoardButton && (
             <StyledExportButton
-              disabled={!(reportData?.boardMetricsCompleted && !reportData?.reportMetricsError?.boardMetricsError)}
+              disabled={!enableExportBoardData}
               onClick={() => handleDownload(REPORT_TYPES.BOARD, startDate, endDate)}
             >
               {COMMON_BUTTONS.EXPORT_BOARD_DATA}
@@ -91,7 +144,7 @@ export const ReportButtonGroup = ({
           )}
           {isShowExportPipelineButton && (
             <StyledExportButton
-              disabled={!!pipelineButtonDisabled}
+              disabled={!enableExportPipelineData}
               onClick={() => handleDownload(REPORT_TYPES.PIPELINE, startDate, endDate)}
             >
               {COMMON_BUTTONS.EXPORT_PIPELINE_DATA}
